Stock: compare and cumulative refactoring (composition).
