+++ ts/Core/Chart/StockChart.ts
@@ -22,7 +22,6 @@ import type {
 } from '../Axis/AxisOptions';
 import type BBoxObject from '../Renderer/BBoxObject';
 import type CSSObject from '../Renderer/CSSObject';
-import type DataExtremesObject from '../Series/DataExtremesObject';
 import type { HTMLDOMElement } from '../Renderer/DOMElementType';
 import type Options from '../Options';
 import type PointerEvent from '../PointerEvent';
@@ -42,30 +41,15 @@ import D from '../DefaultOptions.js';
 const { getOptions } = D;
 import Palette from '../../Core/Color/Palette.js';
 import Point from '../Series/Point.js';
-const {
-    prototype: {
-        tooltipFormatter: pointTooltipFormatter
-    }
-} = Point;
 import Series from '../Series/Series.js';
-const {
-    prototype: {
-        init: seriesInit,
-        processData: seriesProcessData
-    }
-} = Series;
 import SVGRenderer from '../Renderer/SVG/SVGRenderer.js';
 import U from '../Utilities.js';
 const {
     addEvent,
-    arrayMax,
-    arrayMin,
     clamp,
-    correctFloat,
     defined,
     extend,
     find,
-    isArray,
     isNumber,
     isString,
     merge,
@@ -93,13 +77,6 @@ import '../../Extensions/RangeSelector.js';
 declare module '../Axis/AxisLike' {
     interface AxisLike {
         crossLabel?: SVGElement;
-        setCompare(compare?: string, redraw?: boolean): void;
-        setCumulative(cumulative?: boolean, redraw?: boolean): void;
-        setModifier(
-            mode: 'compare'|'cumulative',
-            modeState?: boolean|null|string,
-            redraw?: boolean
-        ): void;
     }
 }
 
@@ -115,34 +92,10 @@ declare module '../Options'{
     }
 }
 
-declare module '../Series/PointLike' {
-    interface PointLike {
-        change?: number;
-        cumulativeSum?: number;
-    }
-}
-
 declare module '../Series/SeriesLike' {
     interface SeriesLike {
         clipBox?: BBoxObject;
-        compareValue?: number;
-        cumulativeTotal?: number;
         forceCropping(): boolean|undefined;
-        getCumulativeExtremes(activeYData: Array<number>): [number, number];
-        initCompare(compare?: string|null): void;
-        initCumulative(cumulative?: boolean): void;
-        modifyValue?(value?: number|null, index?: number): (number|undefined);
-        setCompare(compare?: string|null, redraw?: boolean): void;
-        setCumulative(cumulative?: boolean|null, redraw?: boolean): void;
-    }
-}
-
-declare module '../Series/SeriesOptions' {
-    interface SeriesOptions {
-        compare?: string|null;
-        compareBase?: (0|100);
-        compareStart?: boolean;
-        cumulative?: boolean;
     }
 }
 
@@ -923,239 +876,6 @@ addEvent(Axis, 'afterDrawCrosshair', function (
     });
 });
 
-/* ************************************************************************** *
- *  Start shared compare and cumulative logic                                 *
- * ************************************************************************** */
-
-/**
- * Extend series.init by adding a methods to modify the y values used
- * for plotting on the y axis. For compare mode, this method is called both
- * from the axis when finding dataMin and dataMax,
- * and from the series.translate method.
- *
- * @ignore
- * @function Highcharts.Series#init
- */
-Series.prototype.init = function (): void {
-
-    // Call base method
-    seriesInit.apply(this, arguments as any);
-
-    if (this.options.compare) {
-        // Set comparison mode
-        this.initCompare(this.options.compare);
-    } else if (this.options.cumulative) {
-        // Set Cumulative Sum mode
-        this.initCumulative(this.options.cumulative);
-    }
-};
-
-/**
- * Shared code for the axis.setCompare() and the axis.setCumulative() methods.
- * Inits the 'compare' or the 'cumulative' mode.
- * @private
- */
-Axis.prototype.setModifier = function (
-    mode: 'compare'|'cumulative',
-    modeState?: boolean|null|string,
-    redraw?: boolean
-): void {
-    if (!this.isXAxis) {
-        this.series.forEach(function (series): void {
-            if (
-                mode === 'compare' &&
-                (
-                    isString(modeState) ||
-                    typeof modeState === 'undefined' ||
-                    modeState === null
-                )
-            ) {
-                series.setCompare(modeState, false);
-            } else if (
-                mode === 'cumulative' &&
-                (
-                    typeof modeState === 'boolean' ||
-                    typeof modeState === 'undefined' ||
-                    modeState === null
-                )
-            ) {
-                series.setCumulative(modeState, false);
-            }
-        });
-
-        if (pick(redraw, true)) {
-            this.chart.redraw();
-        }
-    }
-};
-
-// Modify series extremes
-addEvent(
-    Series,
-    'afterGetExtremes',
-    function (e): void {
-        const dataExtremes: DataExtremesObject = (e as any).dataExtremes,
-            activeYData = dataExtremes.activeYData;
-
-        if (this.modifyValue && dataExtremes) {
-            let extremes;
-
-            if (this.options.compare) {
-                extremes = [
-                    this.modifyValue(dataExtremes.dataMin),
-                    this.modifyValue(dataExtremes.dataMax)
-                ];
-            } else if (
-                this.options.cumulative &&
-                isArray(activeYData) &&
-                // If only one y visible, sum doesn't change
-                // so no need to change extremes
-                activeYData.length >= 2
-            ) {
-                extremes = this.getCumulativeExtremes(activeYData);
-            }
-
-            if (extremes) {
-                dataExtremes.dataMin = arrayMin(extremes);
-                dataExtremes.dataMax = arrayMax(extremes);
-            }
-        }
-    }
-);
-
-/**
- * Extend the tooltip formatter by adding support for the point.change variable
- * as well as the changeDecimals option.
- *
- * @ignore
- * @function Highcharts.Point#tooltipFormatter
- *
- * @param {string} pointFormat
- */
-Point.prototype.tooltipFormatter = function (pointFormat: string): string {
-    const point: any = this,
-        { numberFormatter } = point.series.chart,
-        replace = function (value: string): void {
-            pointFormat = pointFormat.replace(
-                '{point.' + value + '}',
-                (point[value] > 0 && value === 'change' ? '+' : '') +
-                    numberFormatter(
-                        point[value],
-                        pick(point.series.tooltipOptions.changeDecimals, 2)
-                    )
-            );
-        };
-
-    if (defined(point.change)) {
-        replace('change');
-    }
-
-    if (defined(point.cumulativeSum)) {
-        replace('cumulativeSum');
-    }
-
-    return pointTooltipFormatter.apply(this, [pointFormat]);
-};
-
-/* ************************************************************************** *
- *  End shared compare and cumulative logic                                   *
- * ************************************************************************** */
-
-/* ************************************************************************** *
- *  Start value compare logic                                                 *
- * ************************************************************************** */
-
-/**
- * Highcharts Stock only. Set the
- * [compare](https://api.highcharts.com/highstock/plotOptions.series.compare)
- * mode of the series after render time. In most cases it is more useful running
- * {@link Axis#setCompare} on the X axis to update all its series.
- *
- * @function Highcharts.Series#setCompare
- *
- * @param {string|null} [compare]
- *        Can be one of `undefined` (default), `null`, `"percent"` or `"value"`.
- *
- * @param {boolean} [redraw=true]
- *        Whether to redraw the chart or to wait for a later call to
- *        {@link Chart#redraw}.
- */
-Series.prototype.setCompare = function (
-    compare?: string|null,
-    redraw?: boolean
-): void {
-    this.initCompare(compare);
-
-    // Survive to export, #5485
-    this.options.compare = this.userOptions.compare = compare;
-
-    if (pick(redraw, true)) {
-        this.chart.redraw();
-    }
-};
-
-/**
- * @ignore
- * @function Highcharts.Series#initCompare
- *
- * @param {string} [compare]
- *        Can be one of `null` (default), `"percent"` or `"value"`.
- */
-Series.prototype.initCompare = function (compare?: string): void {
-    if (compare === 'value' || compare === 'percent') {
-        // Set the modifyValue method
-        this.modifyValue = function (
-            this: Series,
-            value?: number|null,
-            index?: number
-        ): (number|undefined) {
-            if (value === null) {
-                value = 0;
-            }
-
-            const compareValue = this.compareValue;
-
-            if (
-                typeof value !== 'undefined' &&
-                typeof compareValue !== 'undefined'
-            ) { // #2601, #5814
-
-                // Get the modified value
-                if (compare === 'value') {
-                    value -= compareValue;
-                // Compare percent
-                } else {
-                    value = 100 * (value / compareValue) -
-                        (this.options.compareBase === 100 ? 0 : 100);
-                }
-
-                // record for tooltip etc.
-                if (typeof index !== 'undefined') {
-                    const point = this.points[index];
-
-                    if (point) {
-                        point.change = value;
-                    }
-                }
-
-                return value;
-            }
-            return 0;
-        };
-    } else {
-        // When disabling, unset the modifyValue method and clear the points
-        this.modifyValue = null as any;
-        this.points.forEach((point): void => {
-            delete point.change;
-        });
-    }
-
-    // Mark dirty
-    if (this.chart.hasRendered) {
-        this.isDirty = true;
-    }
-};
-
 /**
  * Based on the data grouping options decides whether
  * the data should be cropped while processing.
@@ -1172,235 +892,6 @@ Series.prototype.forceCropping = function (this: Series): (boolean|undefined) {
 
     return groupingEnabled;
 };
-
-/**
- * Extend series.processData by finding the first y value in the plot area,
- * used for comparing the following values
- *
- * @ignore
- * @function Highcharts.Series#processData
- */
-Series.prototype.processData = function (force?: boolean): (boolean|undefined) {
-    let series = this,
-        i,
-        keyIndex = -1,
-        processedXData,
-        processedYData,
-        compareStart = series.options.compareStart === true ? 0 : 1,
-        length,
-        compareValue;
-
-    // call base method
-    seriesProcessData.apply(this, arguments as any);
-
-    if (series.xAxis && series.processedYData) { // not pies
-
-        // local variables
-        processedXData = series.processedXData;
-        processedYData = series.processedYData;
-        length = processedYData.length;
-
-        // For series with more than one value (range, OHLC etc), compare
-        // against close or the pointValKey (#4922, #3112, #9854)
-        if (series.pointArrayMap) {
-            keyIndex = series.pointArrayMap.indexOf(
-                series.options.pointValKey || series.pointValKey || 'y'
-            );
-        }
-
-        // find the first value for comparison
-        for (i = 0; i < length - compareStart; i++) {
-            compareValue = processedYData[i] && keyIndex > -1 ?
-                (processedYData[i] as any)[keyIndex] :
-                processedYData[i];
-            if (
-                isNumber(compareValue) &&
-                (processedXData as any)[i + compareStart] >=
-                (series.xAxis.min as any) &&
-                compareValue !== 0
-            ) {
-                series.compareValue = compareValue;
-                break;
-            }
-        }
-    }
-
-    return;
-};
-
-/**
- * Highcharts Stock only. Set the compare mode on all series
- * belonging to a Y axis.
- *
- * @see [plotOptions.series.compare](https://api.highcharts.com/highstock/plotOptions.series.compare)
- *
- * @sample stock/members/axis-setcompare/
- *         Set compare
- *
- * @function Highcharts.Axis#setCompare
- *
- * @param {string|null} [compare]
- *        The compare mode. Can be one of `undefined` (default), `null`,
- *        `"value"` or `"percent"`.
- *
- * @param {boolean} [redraw=true]
- *        Whether to redraw the chart or to wait for a later call to
- *        {@link Chart#redraw}.
- */
-Axis.prototype.setCompare = function (
-    compare?: string|null,
-    redraw?: boolean
-): void {
-    this.setModifier('compare', compare, redraw);
-};
-
-/* ************************************************************************** *
- *  End value compare logic                                                   *
- * ************************************************************************** */
-
-/* ************************************************************************** *
- *  Start Cumulative Sum logic, author: Rafal Sebestjanski                    *
- * ************************************************************************** */
-
-/**
- * Highcharts Stock only. Set the
- * [cumulative](https://api.highcharts.com/highstock/plotOptions.series.cumulative)
- * mode of the series after render time. In most cases it is more useful running
- * {@link Axis#setCumulative} on the Y axis to update all its series.
- *
- * @function Highcharts.Series#setCumulative
- *
- * @param {boolean} [cumulative=false]
- *        Either enable or disable Cumulative Sum mode.
- *        Can be one of `false` (default) or `true`.
- *
- * @param {boolean} [redraw=true]
- *        Whether to redraw the chart or to wait for a later call to
- *        {@link Chart#redraw}.
- */
-Series.prototype.setCumulative = function (
-    cumulative?: boolean|null,
-    redraw?: boolean
-): void {
-    // Set default value to false
-    cumulative = pick(cumulative, false);
-
-    this.initCumulative(cumulative);
-
-    // Survive to export, #5485
-    this.options.cumulative = this.userOptions.cumulative = cumulative;
-
-    if (pick(redraw, true)) {
-        this.chart.redraw();
-    }
-};
-
-/**
- * @ignore
- * @function Highcharts.Series#initCumulative
- */
-Series.prototype.initCumulative = function (cumulative?: boolean): void {
-    if (cumulative) {
-        // Set the modifyValue method
-        this.modifyValue = function (
-            this: Series,
-            value?: number|null,
-            index?: number
-        ): (number|undefined) {
-            if (value === null) {
-                value = 0;
-            }
-
-            if (value !== void 0 && index !== void 0) {
-                const prevPoint = index > 0 ? this.points[index - 1] : null;
-
-                // Get the modified value
-                if (prevPoint && prevPoint.cumulativeSum) {
-                    value = correctFloat(prevPoint.cumulativeSum + value);
-                }
-
-                // Record for tooltip etc.
-                const point = this.points[index];
-
-                if (point) {
-                    point.cumulativeSum = value;
-                }
-
-                return value;
-            }
-
-            return 0;
-        };
-    } else {
-        // When disabling, unset the modifyValue method and clear the points
-        this.modifyValue = null as any;
-        this.points.forEach((point): void => {
-            delete point.cumulativeSum;
-        });
-    }
-
-    // Mark dirty
-    if (this.chart.hasRendered) {
-        this.isDirty = true;
-    }
-};
-
-/**
- * @ignore
- * @function Highcharts.Series#getCumulativeExtremes
- *
- * @param {Array} [activeYData]
- *        An array cointaining all the points' y values in a visible range.
- */
-Series.prototype.getCumulativeExtremes = function (
-    this: Series,
-    activeYData: Array<number>
-): [number, number] {
-    let cumulativeDataMin = Infinity,
-        cumulativeDataMax = -Infinity;
-
-    activeYData.reduce((prev, cur): number => {
-        const sum = prev + cur;
-
-        cumulativeDataMin = Math.min(cumulativeDataMin, sum, prev);
-        cumulativeDataMax = Math.max(cumulativeDataMax, sum, prev);
-
-        return sum;
-    });
-
-    return [cumulativeDataMin, cumulativeDataMax];
-};
-
-/**
- * Highcharts Stock only. Set the cumulative mode on all series
- * belonging to a Y axis.
- *
- * @see [plotOptions.series.cumulative](https://api.highcharts.com/highstock/plotOptions.series.cumulative)
- *
- * @sample stock/members/axis-setcumulative/
- *         Set cumulative
- *
- * @function Highcharts.Axis#setCumulative
- *
- * @param {boolean} [cumulative]
- *        Whether to disable or enable the cumulative mode.
- *        Can be one of `undefined` (default, treated as `false`),
- *        `false` or `true`.
- *
- * @param {boolean} [redraw=true]
- *        Whether to redraw the chart or to wait for a later call to
- *        {@link Chart#redraw}.
- */
-Axis.prototype.setCumulative = function (
-    cumulative?: boolean,
-    redraw?: boolean
-): void {
-    this.setModifier('cumulative', cumulative, redraw);
-};
-
-/* ************************************************************************** *
- *  End Cumulative Sum logic                                                  *
- * ************************************************************************** */
 
 // Extend the Series prototype to create a separate series clip box. This is
 // related to using multiple panes, and a future pane logic should incorporate
@@ -1503,87 +994,3 @@ addEvent(Chart, 'update', function (
  * */
 
 export default StockChart;
-
-/* *
- *
- *  API Options
- *
- * */
-
-/**
- * Compare the values of the series against the first non-null, non-
- * zero value in the visible range. The y axis will show percentage
- * or absolute change depending on whether `compare` is set to `"percent"`
- * or `"value"`. When this is applied to multiple series, it allows
- * comparing the development of the series against each other. Adds
- * a `change` field to every point object.
- *
- * @see [compareBase](#plotOptions.series.compareBase)
- * @see [Axis.setCompare()](/class-reference/Highcharts.Axis#setCompare)
- * @see [Series.setCompare()](/class-reference/Highcharts.Series#setCompare)
- *
- * @sample {highstock} stock/plotoptions/series-compare-percent/
- *         Percent
- * @sample {highstock} stock/plotoptions/series-compare-value/
- *         Value
- *
- * @type      {string}
- * @since     1.0.1
- * @product   highstock
- * @apioption plotOptions.series.compare
- */
-
-/**
- * Defines if comparison should start from the first point within the visible
- * range or should start from the first point **before** the range.
- *
- * In other words, this flag determines if first point within the visible range
- * will have 0% (`compareStart=true`) or should have been already calculated
- * according to the previous point (`compareStart=false`).
- *
- * @sample {highstock} stock/plotoptions/series-comparestart/
- *         Calculate compare within visible range
- *
- * @type      {boolean}
- * @default   false
- * @since     6.0.0
- * @product   highstock
- * @apioption plotOptions.series.compareStart
- */
-
-/**
- * When [compare](#plotOptions.series.compare) is `percent`, this option
- * dictates whether to use 0 or 100 as the base of comparison.
- *
- * @sample {highstock} stock/plotoptions/series-comparebase/
- *         Compare base is 100
- *
- * @type       {number}
- * @default    0
- * @since      5.0.6
- * @product    highstock
- * @validvalue [0, 100]
- * @apioption  plotOptions.series.compareBase
- */
-
-/**
- * Cumulative Sum feature replaces points' values with the following formula:
- * `sum of all previous points' values + current point's value`.
- * Works only for points in a visible range.
- * Adds the `cumulativeSum` field to each point object that can be accessed
- * e.g. in the [tooltip.pointFormat](https://api.highcharts.com/highstock/tooltip.pointFormat).
- *
- * @see [Axis.setCumulative()](/class-reference/Highcharts.Axis#setCumulative)
- * @see [Series.setCumulative()](/class-reference/Highcharts.Series#setCumulative)
- *
- * @sample {highstock} stock/plotoptions/series-cumulative-sum/
- *         Cumulative Sum
- *
- * @type      {boolean}
- * @default   false
- * @since     next
- * @product   highstock
- * @apioption plotOptions.series.cumulative
- */
-
-''; // keeps doclets above in transpiled file
