+++ ts/Core/Series/Series.ts
@@ -1720,6 +1720,8 @@ class Series {
         series.processedXData = processedData.xData;
         series.processedYData = processedData.yData;
         series.closestPointRange = series.basePointRange = processedData.closestPointRange;
+
+        fireEvent(series, 'afterProcessData');
     }
 
     /**
