+++ ts/masters/modules/stock.src.ts
@@ -12,6 +12,7 @@
 'use strict';
 import Highcharts from '../../Core/Globals.js';
 import OrdinalAxis from '../../Core/Axis/OrdinalAxis.js';
+import CompareCumulativeComposition from '../../Series/CompareCumulativeComposition.js';
 import './broken-axis.src.js';
 import '../../Extensions/DataGrouping.js';
 import '../../Series/HLC/HLCSeries.js';
@@ -30,3 +31,4 @@ G.StockChart = G.stockChart = StockChart.stockChart;
 // Compositions
 Scrollbar.compose(G.Axis);
 OrdinalAxis.compose(G.Axis, G.Series, G.Chart);
+CompareCumulativeComposition.compose(G.Series, G.Axis, G.Point);
